requirements.txt
python-dotenv~=0.21.0
python-telegram-bot==13.15
google-cloud-dialogflow==2.27.*
google-cloud-language==2.12.*
